fix(scene-graph): clean up malformed and negative-indexed binding keys
cleanupStaleBindings could not remove binding keys with non-numeric
index portions (e.g. fills/blendMode) because parseInt returns NaN
and NaN >= length evaluates to false, causing such keys to linger
indefinitely. Also handles negative indices which are never valid.

Split variable-binding-extended test file to stay under 600 lines.
